refactors enum tests
