+++ CedarJava/src/test/java/com/cedarpolicy/SchemaTests.java
@@ -243,7 +243,6 @@ public class SchemaTests {
     class EnumSchemaTests {
 
         @Test
-        @DisplayName("Should parse JSON schema with enum entities")
         void testParseJsonEnumSchema() {
             assertDoesNotThrow(() -> {
                 Schema enumSchema = loadSchemaResource("/enum_schema.json");
@@ -261,36 +260,6 @@ public class SchemaTests {
         }
 
         @Test
-        @DisplayName("Should parse inline enum entity definitions")
-        void testParseInlineEnumDefinitions() {
-            assertDoesNotThrow(() -> {
-                // Test Cedar format inline enum
-                Schema.parse(JsonOrCedar.Cedar, """
-                    entity Color enum ["Red", "Blue", "Green"];
-                    entity Application enum ["TinyTodo"];
-                    """);
-
-                // Test JSON format inline enum
-                Schema.parse(JsonOrCedar.Json, """
-                    {
-                        "": {
-                            "entityTypes": {
-                                "Color": {
-                                    "enum": ["Red", "Blue", "Green"]
-                                },
-                                "Application": {
-                                    "enum": ["TinyTodo"]
-                                }
-                            },
-                            "actions": {}
-                        }
-                    }
-                    """);
-            });
-        }
-
-        @Test
-        @DisplayName("Should reject empty enum definitions")
         void testRejectEmptyEnums() {
             // Test Cedar format empty enum
             assertThrows(Exception.class, () -> {
@@ -300,54 +269,53 @@ public class SchemaTests {
             // Test JSON format empty enum
             assertThrows(Exception.class, () -> {
                 Schema.parse(JsonOrCedar.Json, """
-                    {
-                        "": {
-                            "entityTypes": {
-                                "Color": {
-                                    "enum": []
-                                }
-                            },
-                            "actions": {}
+                        {
+                            "": {
+                                "entityTypes": {
+                                    "Color": {
+                                        "enum": []
+                                    }
+                                },
+                                "actions": {}
+                            }
                         }
-                    }
-                    """);
+                        """);
             });
         }
 
         @Test
-        @DisplayName("Should convert enum schemas between JSON and Cedar formats")
         void testEnumSchemaFormatConversion() throws Exception {
             // Test Cedar to JSON conversion
             Schema cedarEnumSchema = Schema.parse(JsonOrCedar.Cedar, """
-                entity Color enum ["Red", "Blue", "Green"];
-                entity User;
-                action view appliesTo { principal: [User], resource: [User] };
-                """);
+                    entity Color enum ["Red", "Blue", "Green"];
+                    entity User;
+                    action view appliesTo { principal: [User], resource: [User] };
+                    """);
 
             JsonNode jsonResult = cedarEnumSchema.toJsonFormat();
             assertNotNull(jsonResult, "JSON conversion result should not be null");
 
             // Test JSON to Cedar conversion
             String jsonEnumSchema = """
-                {
-                    "": {
-                        "entityTypes": {
-                            "Color": {
-                                "enum": ["Red", "Blue", "Green"]
+                    {
+                        "": {
+                            "entityTypes": {
+                                "Color": {
+                                    "enum": ["Red", "Blue", "Green"]
+                                },
+                                "User": {}
                             },
-                            "User": {}
-                        },
-                        "actions": {
-                            "view": {
-                                "appliesTo": {
-                                    "principalTypes": ["User"],
-                                    "resourceTypes": ["User"]
+                            "actions": {
+                                "view": {
+                                    "appliesTo": {
+                                        "principalTypes": ["User"],
+                                        "resourceTypes": ["User"]
+                                    }
                                 }
                             }
                         }
                     }
-                }
-                """;
+                    """;
             Schema jsonSchemaObj = Schema.parse(JsonOrCedar.Json, jsonEnumSchema);
             String cedarResult = jsonSchemaObj.toCedarFormat();
             assertNotNull(cedarResult, "Cedar conversion result should not be null");
